# app/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fiber-Swagger API",
        "title": "Fiber-Swagger API",
        "contact": {},
        "version": "0.99dev"
    },
    "paths": {
        "/api/Login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoginModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Successful"
                    }
                }
            }
        },
        "/api/Ping": {
            "get": {
                "summary": "Ping Me!",
                "responses": {
                    "200": {
                        "description": "pong"
                    }
                }
            }
        },
        "/api/Secret": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "summary": "Secret",
                "responses": {
                    "200": {
                        "description": "Hello Boss"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get User By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.UserModel"
                        }
                    },
                    "404": {
                        "description": "User Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.LoginModel": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.UserModel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}